added base64-inlining of images
I implemented inlining images as base64 data-uris. I must confess I didn't write tests for it right now (but as I saw, I am not the only one with this problem ;) ).

For the moment the mime type is essentially `image/<anything after the last dot>`, which works for png/jpg, but is for sure a bad hack. It would be better to use [node-mime](https://github.com/broofa/node-mime) and implement it as `'data:' + mime.lookup(src) + '...'`. But this would add an additional dependency which might be a bad thing.